Use bitflags::contains to check for equivalence
Seeing a combination of `==`, `!=`, `is_empty()` and comparison against
`::empty()` which can all be replaced with simple calls to `.contains`.
This function checks if _all_ bitflags in the RHS are contained in LHS
(as opposed to `.intersects` which only checks if there is any flag in
common).

At the same time simplify `memory_types` usage by creating a correctly
sized slice once. This should probably also be used in
`find_memorytype_index`.