Autogenerate README based on library documentation with cargo-readme (#23)
As discussed just last week it's annoying to keep the README and library
docs in sync, in particular because the code in `README.md` isn't
build-tested like the module docs.  Fortunately `cargo-readme` feeds two
birds with one scone: it bases (part of) `README.md` on the module
documentation - which was tested with `cargo test` just prior.