Fix lifetime issues by not calling `.build()` on Vulkan builders
`.build()` discards all lifetime/borrow information attached to the
builder and should be used sparingly - only when a struct is generally
passed by value and does not contain pointer references.

In particular local `&[single_item]` slices get dropped at the end of an
expression, well before the constructed object is passed into a Vulkan
function.  Fortunately creating slices on a single item is as simple as
combining the borrow of that single item with a length of `1` into a
fat-pointer through `std::slice::from_ref`.  When multiple items have to
be passed `.build()` is best called as late as possible, in the same
expression as the Vulkan call receiving the objects.

At the same time `from_ref` helps passing builders into Vulkan functions
(and other builders): builders `Deref` into their Vulkan struct
natrually through `std::slice::from_ref(&some_builder)`, which does not
happen when the builder is moved by-value into a temporary array through
`&[some_builder]`.

In places where it is hard or obnoxious to remove `.build()` - and
structs do not contain pointers - the builder has been replaced with a
direct struct constructor, being more concise while at the same time not
demonstrating this "bad" `.build()` pattern.