vulkan: Use type parameter to safeguard against treating uninitialized memory as `&[u8]`
It is Undefined Behaviour to even cast possibly uninitialized memory as
`&[u8]` until it is fully initialized.

This is a squash of:

Unmark `fn mapped_ptr()` as unsafe and clean up raw pointer casts

Initially coming to the conclusion that `mapped_ptr()` is an `unsafe`
operation in https://github.com/Traverse-Research/gpu-allocator/pull/20
due to the nature of returning a raw pointer (whose lifetime nor
aliasing can be guaranteed), it has been brought to our attention that
this is not inherently `unsafe` and is not marked as such in Rust's
standard library either: take all the `.as_ptr()` methods as precedent.
Interacting with this pointer (dereferencing or converting back to safe
Rust datastructures like slices) employs `unsafe` instead.

Note that this leaves a loophole where Rust can get safe access to
uninitialized memory, which is undefined behaviour.  This will be
addressed in a future changeset.

Use type parameter to safeguard possibly uninitialized mapped memory

Provide SubAllocation::initialize to initalize memory

Safe access to slices holding uninit data

Nightly API for safely initializing MaybeUninit and retrieving init