Unmark `fn mapped_ptr()` as unsafe and clean up raw pointer casts (#32)
Initially coming to the conclusion that `mapped_ptr()` is an `unsafe`
operation in https://github.com/Traverse-Research/gpu-allocator/pull/20
due to the nature of returning a raw pointer (whose lifetime nor
aliasing can be guaranteed), it has been brought to our attention that
this is not inherently `unsafe` and is not marked as such in Rust's
standard library either: take all the `.as_ptr()` methods as precedent.
Interacting with this pointer (dereferencing or converting back to safe
Rust datastructures like slices) employs `unsafe` instead.

Note that this leaves a loophole where Rust can get safe access to
uninitalized memory, which is undefined behaviour.  This will be
addressed in a future changeset.