ci: Include private items in the documentation build-test (#81)
Rust documentation is not only used to document the public API of a
crate, but also as a convenience (ie. intellisense) to developers
working on the crate-internals itself, some of which are not public.
Normally rustdoc would not build these docs but we'd still like to hold
their contents to the same standards (ie. no broken links) as their
public counterpart.