vulkan,d3d12: Remove unused `backtrace` field from `Allocator`
Only `SubAllocator`s print their backtrace and need to hold on to this
field; there is no need for the `Allocator` to hold on to and own the
exact same value.

Rust 1.57 now detects such unused fields, when they have previously only
been used inside `derive()` macros such as `Debug`.