Upgrade to ash 0.34 with ash-window 0.8 (#87)
This ash release switches over to a dynamically linked scheme where
Vulkan is linked directly instead of through libloading by default.  As
the main library crate doesn't need to load an entrypoint such linking
is not desired at all and is disabled by default (as is runtime linking,
effectively no entrypoints can be loaded): it must be re-enabled by hand
when building the examples.