Simplify example folder structure
Examples are intended to be single files or modules at best, not
separate "crate-like" directories with a `src/` subdirectory yet without
`Cargo.toml`.  Not only is this [package layout documented], it also
removes the need for a specific `path =` setup as `cargo` is now able to
relate the `name` of every `[[example]]` to a specific file or directory
(`examples/<name>.rs` or `examples/<name>/main.rs`).

[package layout documented]: https://doc.rust-lang.org/cargo/guide/project-layout.html