vulkan: Use `Send`+`Sync` pointer wrapper type to auto-derive these marker traits on `MemoryBlock`/`Allocation` (#153)
Instead of manually unsafely implementing these markers on the
aforementioned structs, and potentially wrongly leaving them in place
while adding a field that is not `Send`/`Sync` safe, define these unsafe
markers on a wrapper type around `mapped_ptr` exclusively.  This allows
Rust to auto-implement the markers for the structs currently, because
all other fields are (fortunately!) `Send`+`Sync`.