examples: Initialize `env_logger` to actually see `log` output (#145)
`log` only provides the facade, and needs an explicitly initialized
logger to emit output to `stdout`.  The `env_logger` crate is typically
used across the Rust ecosystem, but unfortunately initializes with a
default of `off` unless the `RUST_LOG` variable is set: default it to
`trace` instead (we mostly log on `info` though).