CI: Replace unnecessary `actions-rs` steps with direct `run: cargo ...`
GitHub Actions' `runner-images` already come fully loaded with all the
Rust tools that we need [1]: remove the bloated and outdated
`actions-rs/toolchain` setup step for that, and also remove the
`actions-rs/cargo` action that only serves as a more convoluted way to
run simple `cargo` shell commands.

Besides, `actions-rs` are unmaintained and outdated, and throw an
alarming number of spammy warnings in CI build output.

[1]: https://github.com/actions/runner-images/blob/main/images/linux/Ubuntu2204-Readme.md#rust-tools