Make `metal` dependency rely on the `metal` feature
We seem to have this weird case where the `metal` crate is always
included (on supported target OSes), even if the non-default `metal`
feature isn't turned on.  Found this while wondering why `cargo` wasn't
complaining about a clash thanks to the implicit feature created by an
optional crate with an identically-named feature in the `[features]`
table without [the `dep:` rename prefix], turns out that the `metal`
crate wasn't `optional = true` in the first place.  Fix that.

[the `dep:` rename prefix]: https://doc.rust-lang.org/cargo/reference/features.html#optional-dependencies