Avoid `<=` dependency version specifications that may break builds.
`<=` version requirements may break builds in the future.
In particular, if `egui` releases a version `0.27.1`, or `windows`
releases a version `0.58.1`, and another package in the dependency graph
requires one of those versions, Cargo will report a fatal error. This is
because  `<=0.58` does not mean “less than or equal to any version in
`0.58.*`”, it means “less than or equal to 0.58.0” — the third component
is implicitly zero — and Cargo will not allow two semver-compatible
versions of the same package in the dependency graph.

> Avoid constraining the upper bound of a version to be anything less
> than the next semver incompatible version (e.g. avoid ">=2.0, <2.4")
> as other packages in the dependency tree may require a newer version,
> leading to an unresolvable error (see #9029)

— https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html#multiple-version-requirements

This change avoids this problem by using only `<` requirements.
I recommend following this rule in future changes too.