cargo: Bump `windows` crate range to `0.53-0.59` (#258)
* cargo: Bump `windows` crate range to `0.53-0.59`

Note that `windows 0.59` bumps the MSRV to `1.74`, but we still support
this crate dependency all the way back to `0.53`.

* CI: Use minimal dependency versions when running MSRV check

For the time being we have to instruct `cargo` to use some lower version
of our dependencies to ensure our MSRV doesn't need to be bumped -
that is one of the resons why we have a supported version range on our
dependencies.  In the future when Rust 1.84 becomes our MSRV, it will
be able to resolve these automatically (assumming crates set their
`rust-version` correctly if at all!):
https://blog.rust-lang.org/2025/01/09/Rust-1.84.0.html#cargo-considers-rust-versions-for-dependency-version-selection

* Address new clippy lints