Upgrade to `objc2 0.6` and `objc2-metal 0.3` (#261)
* Upgrade to `objc2 0.6` and `objc2-metal 0.3`

The `MTLCreateSystemDefaultDevice()` function finally has a high-level
wrapper that constructs a `Retained` protocol object for us.

Note also that a new `MTLAllocation` interface was added to Metal.  Even
if that interface is _optional_, current `objc2` design requires this
trait to be part of the hierarchy (and the feature explicitly enabled)
to get access to all "descendant" interfaces (`MTLResource`, which
provides `MTLBuffer` and `MTLTexture`).

* Bump MSRV to 1.71 for `objc2`