Add reserved_bytes method for Allocator (#266)
* Change total_reserved_bytes to total_capacity_bytes in AllocatorReport

The use of "reserved" may be confused with the concept of reserved
memory in Dx12. Capacity is in line with the similar concept from (e.g.)
std::vec::Vec.

* Add capacity method for Allocator

This fills the use case where one needs the total capacity given by the
sum of sizes of all device-allocated memory blocks, but compiling a
complete report (in particular due to the calls to report_allocations
from the SubAllocator) would be too costly.