Remove stale and UB `winapi` compatibility layer (#256)
With the `windows` migration we introduced a `winapi` compatibility
layer to make it easier to use `winapi` types with `gpu-allocator`.  As
most of the ecosystem has switched to `windows` it is now time to sunset
these helpers which are otherwise tedious to test and maintain, not to
mention contain long(er) standing unsoundness bugs.